refactor common fns into utils
