+++ scripts/wfl/twisttcap/find_running_wfl_workloads.py
@@ -1,27 +1,10 @@
 import argparse
 import json
 import requests
-import google.auth as googleauth
 from firecloud import api as fapi
 from time import sleep
 from pprint import pprint
-
-
-def get_access_token():
-    """Get access token."""
-    creds, _ = googleauth.default()
-    auth_req = googleauth.transport.requests.Request()
-    creds.refresh(auth_req)
-
-    return creds.token
-
-def get_headers(request_type='get'):
-    headers = {"Authorization": "Bearer " + get_access_token(),
-                "accept": "application/json"}
-    if request_type == 'post':
-        headers["Content-Type"] = "application/json"
-    return headers
-
+from utils import get_headers
 
 
 def retrieve_WFL_workloads(wfl_project):
